test(arel): move 297 TS-only tests out of Rails-named files (RFC 0124)
`pnpm parity:test` reported arel at 707/707 with 297 "extra (TS only)"
tests, all of them living inside the Rails-named test files rather than
the `.trails.test.ts` twins the package keeps for exactly this purpose.

Every `it()` in a Rails-named arel test file now has a same-named Rails
test:

- 297 TS-only tests move to the sibling `.trails.test.ts` (eight created
  where none existed), bodies and names unchanged. Shared `const`s,
  helpers and the comments that document each test travel with them.
- Two loop-generated groups in `visitors/to-sql.test.ts`
  (`describe("value-class visitors aliased to unsupported")`) move too —
  their template-literal names are invisible to the extractor, so they
  never showed up in the extra count.
- The nine unsuffixed TS-only files are renamed `*.trails.test.ts`;
  `predications.test.ts` folds into its existing twin.
- `collectors/sql-string.test.ts` "returned sql uses utf8 encoding" now
  compiles `astWithBinds()` as Rails does
  (collectors/sql_string_test.rb:35-38) and asserts a strict UTF-8 round
  trip — the nearest TS-expressible form of `sql.encoding`. It was
  building its own collector and asserting `typeof === "string"`.
- The two `.trails` tests that reused a Rails test name verbatim
  (`delete-manager` "handles limit properly", `visitors/sqlite`
  "does not support boolean") are renamed to what they actually assert.

arel stays 707/707 with 0 extra and 0 wrong-describe; the suite is 1447
tests before and after.

Also: CLAUDE.md's zero-import-slot inventory said "two instances exist
and are the only ones" while `arel/src/node-slots.ts` — the largest of
the three, and one that cites that very section as its justification —
went unlisted. An inventory that omits an instance weakens the rule it
polices.

Closes-story: arel-trails-only-tests-in-rails-named-files
Closes-story: claude-md-slot-inventory-omits-arel-node-slots
