+++ packages/arel/src/nodes/not.test.ts
@@ -7,13 +7,6 @@ describe("Arel", () => {
   const users = new Table("users");
 
   describe("not", () => {
-    it("makes a NOT node", () => {
-      const eq = users.get("id").eq(1);
-      const not = new Nodes.Not(eq);
-      expect(not).toBeInstanceOf(Nodes.Not);
-      expect(not.expr).toBe(eq);
-    });
-
     describe("equality", () => {
       it("is equal with equal ivars", () => {
         const array = [
